move extension functions into packages with the same name as the stuff they are extending
