+++ src/test/kotlin/io/inbot/eskotlinwrapper/AbstractElasticSearchTest.kt
@@ -2,6 +2,7 @@ package io.inbot.eskotlinwrapper
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.elasticsearch.client.RestHighLevelClient
+import org.elasticsearch.client.crudDao
 import org.elasticsearch.common.xcontent.XContentType
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
